requirements.txt
discord.py[voice]>=2.0
yt-dlp
python-dotenv
pytest
